Wind Updates
Changed all wind items to use the same function.
Damage is now in 2 parts: Base and Bonus.
Base Damage is 50% of the total damage.
Bonus Damage is variable by Doubled status (50% bonus for doubled per weapon) and by current HP. 0-50% HP Gives 0% Bonus Damage, and above it scales linearly. 
The whole damage value varies based on distance as it used to. 

The ultimate uses the same function now. The old one is still in place, but it felt silly to duplicate this functionality. There is a big bit of the WindUlt which may have been necessary about checking if the caster unit is a hero? Not sure what's going on with that. We might need to relook.